delete sanitizer, move logic to respective extensions
